Make it possible to not build DSSI plugins.
If HAVE_LIBLO wasn't true, then DSSI plugins would (mostly) not be
built, but the install rule would still try to install them and fail.

Add a BUILD_DSSI variable, so the test for whether DSSI is supported is
done in one place, and check it in the install rule.